% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{add_ctype_data_to_container}
\alias{add_ctype_data_to_container}
\title{Add an environment for a cell type to the container}
\usage{
add_ctype_data_to_container(container, scMinimal)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{scMinimal}{environment A sub-container for the project with one
cell type's gene expression data in its raw and processed forms as well
as metadata}
}
\value{
the container with added scMinimal for a cell type in a slot under
container$scMinimal_ctypes
}
\description{
Add an environment for a cell type to the container
}
