% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctype_vargenes.R
\name{apply_cellwise_batch_correct}
\alias{apply_cellwise_batch_correct}
\title{Run combat on donor mean expression matrix}
\usage{
apply_cellwise_batch_correct(container, batch_var)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{batch_var}{character The name of the metadata variable corresponding
with the batches to remove}
}
\value{
the project container with corrected matrices in 
scMinimal_ctype[[ct]]$data_means
}
\description{
Run combat on donor mean expression matrix
}
