% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{apply_combat}
\alias{apply_combat}
\title{Apply ComBat batch correction}
\usage{
apply_combat(container, batch_var)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{batch_var}{character A batch variable from metadata to remove}
}
\value{
the project container with pseudobulked matrices corrected for batch
}
\description{
Apply ComBat batch correction
}
