% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{check_for_all_null}
\alias{check_for_all_null}
\title{Checks if there are any non-NULL elements in a list}
\usage{
check_for_all_null(mylist)
}
\arguments{
\item{mylist}{list The list to check}
}
\value{
TRUE if there exist non-NULL elements or FALSE if all are NULL
}
\description{
Checks if there are any non-NULL elements in a list
}
