% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{clean_data}
\alias{clean_data}
\title{Clean data to remove genes only expressed in a few cells and donors
with very few cells. Generally, this should be done through calling the
form_tensor() wrapper function.}
\usage{
clean_data(container, donor_min_cells = 5)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{donor_min_cells}{numeric Minimum threshold for number of cells per
donor (default=5)}
}
\value{
The project container with cleaned counts matrices in each 
container$scMinimal_ctype$<ctype>$count_data.
}
\description{
Clean data to remove genes only expressed in a few cells and donors
with very few cells. Generally, this should be done through calling the
form_tensor() wrapper function.
}
