% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_by_donors.R
\name{collapse_by_donors}
\alias{collapse_by_donors}
\title{Computes mean expression for each gene in each cell type per donor}
\usage{
collapse_by_donors(container, shuffle = F)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{shuffle}{logical If TRUE shuffles cell to donor linkages (default=FALSE)}
}
\value{
an updated container with the collapsed gene by donor average matrices
in the slots of container$scMinimal_ctype
}
\description{
Computes mean expression for each gene in each cell type per donor
}
