% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colMeanVars}
\alias{colMeanVars}
\title{calculate column mean and variance, optionally taking a subset of rows to operate on}
\usage{
colMeanVars(sY, rowSel, ncores = 1L)
}
\arguments{
\item{sY}{sparse matrix Gene by cell matrix of counts}

\item{rowSel}{numeric The selected rows (genes)}

\item{ncores}{numeric The number of cores}
}
\description{
calculate column mean and variance, optionally taking a subset of rows to operate on
}
