% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_factors.R
\name{compare_factors}
\alias{compare_factors}
\title{Compare gene loadings between any two factors}
\usage{
compare_factors(
  container,
  f_compare,
  direction,
  compare_type,
  sig_thresh = 0.05
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{f_compare}{numeric The number of the two factors to compare}

\item{direction}{character Vector specifying the directions to compare for each
factor. Can either be 'up' or 'down' for each factor.}

\item{compare_type}{character Set to either 'same' or 'different' to get
genes that are significant in both or only significant in one or the other,
respectively.}

\item{sig_thresh}{numeric The significance p-value threshold to use (default=0.05)}
}
\value{
The comparison heatmap in container$plots$comparisons$<f1>_<f2>
}
\description{
Compare gene loadings between any two factors
}
