% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{compute_donor_props}
\alias{compute_donor_props}
\title{Get donor proportions of each cell type or subtype}
\usage{
compute_donor_props(clusts, metadata)
}
\arguments{
\item{clusts}{integer Cluster assignments for each cell with names as cell barcodes}

\item{metadata}{data.frame The $metadata field for the given scMinimal}
}
\value{
a data.frame of cluster proportions for each donor
}
\description{
Get donor proportions of each cell type or subtype
}
