% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LR_interact.R
\name{compute_LR_interact}
\alias{compute_LR_interact}
\title{Plot the LR interactions for one factor}
\usage{
compute_LR_interact(
  container,
  lr_pairs,
  factor_select,
  sig_thresh = 0.05,
  percentile_exp_rec = 0.75
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{lr_pairs}{data.frame Data of ligand-receptor pairs. First column should
be ligands and second column should be one or more receptors separated by an
underscore such as receptor1_receptor2 in the case that multiple receptors are
required for signaling.}

\item{factor_select}{numeric The factor to get associated LR-modules for}

\item{sig_thresh}{numeric The p-value significance threshold to use for module-
factor associations and ligand-factor associations}
}
\value{
The project container with the plt added in container$plots$lr_analysis$factor_num
}
\description{
Plot the LR interactions for one factor
}
