% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{count_word}
\alias{count_word}
\title{count_word. From older version of simplifyEnrichment package.}
\usage{
count_word(term, exclude_words = NULL)
}
\arguments{
\item{term}{A vector of description texts.}

\item{exclude_words}{The words that should be excluded.}
}
\value{
A data frame with words and frequencies.
}
\description{
count_word. From older version of simplifyEnrichment package.
}
