% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{determine_ranks_tucker}
\alias{determine_ranks_tucker}
\title{Rank determination method}
\usage{
determine_ranks_tucker(
  container,
  max_ranks_test,
  method = "svd",
  num_iter = 100
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{max_ranks_test}{numeric Vector of length 3 with maximum number of
ranks to test for donor, gene, and cell type modes in that order}

\item{method}{character The method to use for rank determination. For 'svd',
reconstruction errors are computed using svd on the unfolded tensor along each
mode. For 'tucker', reconstruction errors are computed using tucker decomposition
(default='svd')}

\item{num_iter}{numeric Number of null iterations (default=100)}
}
\value{
the project container with rank determination plot and results
added into slots
}
\description{
Rank determination method
}
