% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature.names}
\alias{feature.names}
\title{Ensemble to gene symbol mapping}
\format{
An object of class \code{data.frame} with 32738 rows and 2 columns.
}
\usage{
feature.names
}
\description{
Ensemble to gene symbol mapping
}
\keyword{datasets}
