% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{form_tensor}
\alias{form_tensor}
\title{Form the tensor and scale the data}
\usage{
form_tensor(container, var_scale_power = NULL, batch_var = NULL)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{var_scale_power}{numeric Exponent of normalized variance that is
used for variance scaling. Variance for each gene
is initially set to unit variance across donors (for a given cell type).
Variance for each gene is then scaled by multiplying the unit scaled values
by each gene's normalized variance (where the effect of the mean-variance
dependence is taken into account) to the exponent specified here.
If NULL, uses var_scale_power from container$experiment_params. (default=NULL)}

\item{batch_var}{character A batch variable from metadata to remove (default=NULL)}
}
\value{
the project container with tensor data added in the
container$tensor_data slot
}
\description{
Form the tensor and scale the data
}
