% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{get_all_lds_factor_plots}
\alias{get_all_lds_factor_plots}
\title{Generate loadings heatmaps for all factors}
\usage{
get_all_lds_factor_plots(
  container,
  use_sig_only = FALSE,
  annot = "none",
  pathways_list = NULL,
  sig_thresh = 0.05,
  display_genes = FALSE
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{use_sig_only}{logical If TRUE, includes only significant genes
from jackstraw in the heatmap. If FALSE, includes all the variable genes.
(default = FALSE)}

\item{annot}{character If set to "pathways" then creates an adjacent heatmap
showing which genes are in which pathways. If set to "sig_genes" then creates
an adjacent heatmap showing which genes were significant from jackstraw. If
set to "none" no adjacent heatmap is plotted. (default="none")}

\item{pathways_list}{list A list of sets of pathways for each factor. List index
should be the number corresponding to the factor. (default=NULL)}

\item{sig_thresh}{numeric Pvalue significance threshold to use. If use_sig_only is
TRUE the threshold is used as a cutoff for genes to include. If annot is "sig_genes"
this value is used in the gene significance colormap as a minimum threshold. (default=0.05)}

\item{display_genes}{logical If TRUE, displays the names of gene names (default=FALSE)}
}
\value{
the project container with the list of plots placed in container$plots$all_lds_plots
}
\description{
Generate loadings heatmaps for all factors
}
