% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_all_subclust_plots}
\alias{get_all_subclust_plots}
\title{Generate subcluster plots for multiple subclusterings at different resolutsions}
\usage{
get_all_subclust_plots(container, ctypes, res, factors)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctypes}{character The major cell types for which to get subcluster plots for}

\item{res}{numeric The subcluster resolution corresponding to the ctypes vector}

\item{factors}{numeric A vector of factors to get compute subtype associations
with. Should be same length as ctypes vector.}
}
\value{
the subtype plots for cell types at the specified resolutions
}
\description{
Generate subcluster plots for multiple subclusterings at different resolutsions
}
