% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_analysis.R
\name{get_avmax_correlations}
\alias{get_avmax_correlations}
\title{Compute correlations between factors from full dataset decomposition to those of
from the subsampled dataset decomposition}
\usage{
get_avmax_correlations(res_full, res_sub, res_use)
}
\arguments{
\item{res_full}{matrix Either loadings or donor scores matrix from
Tucker decomposition on full dataset}

\item{res_sub}{matrix Either loadings or donor scores matrix from
Tucker decomposition on subsampled dataset}

\item{res_use}{character Set to either 'loadings' or 'dscores' to indicate
the data type that has been entered}
}
\value{
the average of all maximum correlations between factors of the two decompositions
}
\description{
Compute correlations between factors from full dataset decomposition to those of
from the subsampled dataset decomposition
}
