% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctype_vargenes.R
\name{get_ctype_data}
\alias{get_ctype_data}
\title{Create an environment for each cell type}
\usage{
get_ctype_data(
  container,
  make_clean = TRUE,
  donor_min_cells = 5,
  gene_min_cells = 5
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{make_clean}{logical TRUE to apply minimum thresholds for number of cells
expressing a gene and number of cells per donor (default=TRUE)}

\item{donor_min_cells}{numeric Minimum threshold for number of cells per
donor. Only used if make_clean = TRUE. (default=5)}

\item{gene_min_cells}{numeric Minimum threshold for number of cells
with nonzero expression of a gene. Only used if make_clean = TRUE. (default=5)}
}
\value{
the project container with the scMinimal environments added into
the container$scMinimal_ctypes slot
}
\description{
Create an environment for each cell type
}
