% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_ctype_prop_associations}
\alias{get_ctype_prop_associations}
\title{Compute and plot associations between donor factor scores and donor proportions of major cell types}
\usage{
get_ctype_prop_associations(container, stat_type, n_col = 2)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{stat_type}{character Either "fstat" to get F-Statistics, "adj_rsq" to get adjusted
R-squared values, or "adj_pval" to get adjusted pvalues.}

\item{n_col}{numeric The number of columns to organize the plots into (default=2)}
}
\value{
The project container with a cowplot figure of results plots in 
container$plots$ctype_prop_factor_associations.
}
\description{
Compute and plot associations between donor factor scores and donor proportions of major cell types
}
