% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_analysis.R
\name{get_dnr_avmax_correlations}
\alias{get_dnr_avmax_correlations}
\title{Compute correlations between donor scores from full dataset decomposition to those of
the donor scores from the subsampled dataset decomposition}
\usage{
get_dnr_avmax_correlations(dnr_full, dnr_sub)
}
\arguments{
\item{dnr_full}{matrix Donor scores matrix from Tucker decomposition on full dataset}

\item{dnr_sub}{matrix Donor scores matrix from Tucker decomposition on subsampled dataset}
}
\value{
the average of all maximum correlations between factors of the two decompositions
}
\description{
Compute correlations between donor scores from full dataset decomposition to those of
the donor scores from the subsampled dataset decomposition
}
