% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{get_donor_meta}
\alias{get_donor_meta}
\title{Gets core donor metadata with columns of donors as minimum. Will subset the matrix
to only donors processed in the analysis unless otherwise specified.}
\usage{
get_donor_meta(container, additional_meta = NULL, only_analyzed = TRUE)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{additional_meta}{character A vector of other variables to include (default=NULL)}

\item{only_analyzed}{logical Set to TRUE to only include donors that were included
in the formed tensor, otherwise set to FALSE (default=TRUE)}
}
\value{
the container with the minimal metadata in container$donor_metadata
}
\description{
Gets core donor metadata with columns of donors as minimum. Will subset the matrix
to only donors processed in the analysis unless otherwise specified.
}
