% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{get_explained_var}
\alias{get_explained_var}
\title{Get explained variance for each cell type for one factor}
\usage{
get_explained_var(container, tmp_casted_num, factor_use)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{tmp_casted_num}{matrix The gene by cell type loadings matrix}

\item{factor_use}{numeric The factor to investigate}
}
\value{
explained variance for each cell type in a list
}
\description{
Get explained variance for each cell type for one factor
}
