% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{get_factor_combos}
\alias{get_factor_combos}
\title{Gets valid combinations of numbers of ranks to test with tucker}
\usage{
get_factor_combos(max_ranks_test)
}
\arguments{
\item{max_ranks_test}{numeric Vector of length 3 with maximum number of
ranks to test for donor, gene, and cell type modes in that order}
}
\value{
a dataframe containing donor, gene, cell type rank combinations
}
\description{
Gets valid combinations of numbers of ranks to test with tucker
}
