% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{get_fstats_pvals}
\alias{get_fstats_pvals}
\title{Calculate adjusted p-values for gene_celltype fiber-donor score associations}
\usage{
get_fstats_pvals(fstats_real, fstats_shuffled)
}
\arguments{
\item{fstats_real}{numeric A vector of F-Statistics for gene-cell type-factor combinations}

\item{fstats_shuffled}{numeric A vector of null F-Statistics}
}
\value{
A vector of adjusted p-values for associations of the unshuffled fibers with
factor donor scores.
}
\description{
Calculate adjusted p-values for gene_celltype fiber-donor score associations
}
