% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw_v2.R
\name{get_fstats_pvals_v2}
\alias{get_fstats_pvals_v2}
\title{Calculate adjusted pvalues for gene celltype fiber-donor score associations}
\usage{
get_fstats_pvals_v2(fstats_real, fstats_shuffled_all)
}
\arguments{
\item{fstats_real}{numeric A vector of F-Statistics for gene-cell type-factor combinations}

\item{fstats_shuffled}{numeric A vector of null F-Statistics}
}
\value{
adjusted pvalues for associations of the unshuffled fibers with factor donor scores
}
\description{
Calculate adjusted pvalues for gene celltype fiber-donor score associations
}
