% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LR_interact.R
\name{get_gene_modules}
\alias{get_gene_modules}
\title{Compute WGCNA gene modules for each cell type}
\usage{
get_gene_modules(container, sft_thresh)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{sft_thresh}{numeric A vector indicating the soft threshold to use for
each cell type. Length should be the same as container$experiment_params$ctypes_use}
}
\value{
The project container with gene modules added
}
\description{
Compute WGCNA gene modules for each cell type
}
