% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{get_gene_set_vectors}
\alias{get_gene_set_vectors}
\title{Get logical vectors indicating which genes are in which pathways}
\usage{
get_gene_set_vectors(container, gene_sets, tmp_casted_num)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{gene_sets}{character Vector of gene sets to extract genes for}

\item{tmp_casted_num}{matrix The gene by cell type loadings matrix}
}
\value{
A list of the logical vectors for each pathway.
}
\description{
Get logical vectors indicating which genes are in which pathways
}
