% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{get_hybrid_sim}
\alias{get_hybrid_sim}
\title{Add batch effect to real dataset}
\usage{
get_hybrid_sim(container, counts = NULL, nbatches = 3, batch_override = NULL)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses (default=NULL)}

\item{counts}{matrix Gene by cell matrix of counts. Use this if want to add batch at
count level. (default=NULL)}

\item{nbatches}{numeric The number of batches to create. Donors are randomly
assigned to batches. (default=3)}

\item{batch_override}{list A list of vectors soecifying donors in each batch
(default=NULL)}
}
\value{
either project container with batch added to normalized data or the counts
matrix with batch effect added to it
}
\description{
Add batch effect to real dataset
}
