% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{get_hybrid_sim_nonlin}
\alias{get_hybrid_sim_nonlin}
\title{Add batch effect to real dataset}
\usage{
get_hybrid_sim_nonlin(container, nbatches = 2)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses (default=NULL)}

\item{nbatches}{numeric The number of batches to create. Donors are randomly
assigned to batches. (default=2)}
}
\value{
either project container with batch added to normalized data or the counts
matrix with batch effect added to it
}
\description{
Add batch effect to real dataset
}
