% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_indv_subtype_associations}
\alias{get_indv_subtype_associations}
\title{Compute single subtype associations with a factor}
\usage{
get_indv_subtype_associations(container, donor_props, factor_select)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{donor_props}{matrix Donor proportions of subtypes}

\item{factor_select}{numeric The factor to get associations for}
}
\value{
the association statistics for each factor
}
\description{
Compute single subtype associations with a factor
}
