% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_analysis.R
\name{get_lds_avmax_correlations}
\alias{get_lds_avmax_correlations}
\title{Compute correlations between loadings from full dataset decomposition to those of
the loadings from the subsampled dataset decomposition}
\usage{
get_lds_avmax_correlations(lds_full, lds_sub)
}
\arguments{
\item{lds_full}{matrix Loadings matrix from Tucker decomposition on full dataset}

\item{lds_sub}{matrix Loadings matrix from Tucker decomposition on subsampled dataset}
}
\value{
the average of all maximum correlations between factors of the two decompositions
}
\description{
Compute correlations between loadings from full dataset decomposition to those of
the loadings from the subsampled dataset decomposition
}
