% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stability.R
\name{get_max_correlations}
\alias{get_max_correlations}
\title{Computes the max correlation between each factor of the decomposition done using
the whole dataset to each factor computed using the subsampled/bootstrapped dataset}
\usage{
get_max_correlations(res_full, res_sub, res_use)
}
\arguments{
\item{res_full}{matrix Either the donor scores or loadings matrix from the original
decomposition}

\item{res_sub}{matrix Either the donor scores or loadings matrix from the new
decomposition}

\item{res_use}{character Can either be 'loadings' or 'dscores' and should correspond
with the data matrix used}
}
\value{
a vector of the max correlations for each original factor
}
\description{
Computes the max correlation between each factor of the decomposition done using
the whole dataset to each factor computed using the subsampled/bootstrapped dataset
}
