% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_means}
\alias{get_means}
\title{calculates factor-stratified means for each column}
\usage{
get_means(sY, rowSel, numCells)
}
\arguments{
\item{sY}{sparse matrix Gene by cell matrix of counts}

\item{rowSel}{factor The donor that each cell is from}

\item{numCells}{numeric Total number of cells belonging to each donor}
}
\description{
calculates factor-stratified means for each column
}
