% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized_variance.R
\name{get_normalized_variance}
\alias{get_normalized_variance}
\title{Get normalized variance for each gene. This takes into account the
mean-variance relation. Based on a Pagoda2 method...}
\usage{
get_normalized_variance(scMinimal, use_counts = F)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms as well
as metadata}

\item{use_counts}{logical If TRUE, uses raw expression counts for computing
the normalized variance. If FALSEd, uses the normalized, log-transformed
values. (default=FALSE)}
}
\value{
a vector of the normalized variance for each gene
}
\description{
Get normalized variance for each gene. This takes into account the
mean-variance relation. Based on a Pagoda2 method...
}
