% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{get_num_batch_ranks}
\alias{get_num_batch_ranks}
\title{Get plot of r-squared values for each factor's association with a batch variable
to varying numbers of donor factors}
\usage{
get_num_batch_ranks(container, donor_ranks_test, batch_var, thresh = 0.5)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{donor_ranks_test}{numeric The number of donor rank values to test}

\item{batch_var}{character The name of the batch meta variable}

\item{thresh}{numeric The threshold r-squared cutoff for considering a
factor to be a batch factor. Can be a vector of multiple values to get plots
at varying thresholds. (default=0.5)}
}
\value{
plots of factor-batch associations for varying numbers of factors
}
\description{
Get plot of r-squared values for each factor's association with a batch variable
to varying numbers of donor factors
}
