% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{get_one_factor_gene_pvals}
\alias{get_one_factor_gene_pvals}
\title{Get significant genes for a factor}
\usage{
get_one_factor_gene_pvals(container, factor_select)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The number corresponding to the factor to extract}
}
\value{
A gene by cell type matrix of gene significance p-values for a factor
}
\description{
Get significant genes for a factor
}
