% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{get_real_fstats}
\alias{get_real_fstats}
\title{Get F-Statistics for the real (non-shuffled) gene_ctype fibers}
\usage{
get_real_fstats(container, ncores)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ncores}{numeric The number of cores to use}
}
\value{
a vector F-Statistics for each gene-celltype-factor combination
}
\description{
Get F-Statistics for the real (non-shuffled) gene_ctype fibers
}
