% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{get_reconstruct_errors_tucker}
\alias{get_reconstruct_errors_tucker}
\title{Calculate reconstruction errors using Tucker approach}
\usage{
get_reconstruct_errors_tucker(tnsr, max_ranks_test, ncores)
}
\arguments{
\item{tnsr}{array A 3-dimensional array with dimensions of
donors, genes, and cell types in that order}

\item{max_ranks_test}{numeric Vector of length 3 with maximum number of
ranks to test for donor, gene, and cell type modes in that order}

\item{ncores}{numeric The number of cores to use}
}
\value{
the reconstruction errors
}
\description{
Calculate reconstruction errors using Tucker approach
}
