% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_de_hmaps}
\alias{get_subclust_de_hmaps}
\title{Get list of cell subtype differential expression heatmaps}
\usage{
get_subclust_de_hmaps(container, all_ctypes, all_res)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{all_ctypes}{character A vector of the cell types to include}

\item{all_res}{numeric A vector of resolutions matching the all_ctypes parameter}
}
\value{
A list of cell subcluster DE marker gene heatmaps as grob objects.
}
\description{
Get list of cell subtype differential expression heatmaps
}
