% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_enr}
\alias{get_subclust_enr}
\title{Title}
\usage{
get_subclust_enr(container, all_ctypes, all_res, all_factors)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{all_ctypes}{character A vector of the cell types to include}

\item{all_res}{numeric A vector of resolutions matching the all_ctypes parameter}

\item{all_factors}{numerc A vector of the factors to compute associations for}
}
\value{
the association heatmap object
}
\description{
Title
}
