% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_enr_bplot}
\alias{get_subclust_enr_bplot}
\title{Get barplot showing significance of associations for cell subtypes}
\usage{
get_subclust_enr_bplot(container, ctype, factor_use)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctype}{character The cell type to plot}

\item{factor_use}{numeric The factor to plot}
}
\value{
the plot in container$plots$subc_bplots$<ctype>
}
\description{
Get barplot showing significance of associations for cell subtypes
}
