% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_enr_fig}
\alias{get_subclust_enr_fig}
\title{Get a figure showing cell subtype proportion associations with each factor. Combines
this plot with subtype UMAPs and differential expression heatmaps. Note that this
function runs better if the number of cores in the conos object in
container$embedding has n.cores set to a relatively small value < 10.}
\usage{
get_subclust_enr_fig(container, all_ctypes, all_res)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{all_ctypes}{character A vector of the cell types to include}

\item{all_res}{numeric A vector of resolutions matching the all_ctypes parameter}
}
\value{
A cowplot figure placed in the slot container$plots$subc_fig.
}
\description{
Get a figure showing cell subtype proportion associations with each factor. Combines
this plot with subtype UMAPs and differential expression heatmaps. Note that this
function runs better if the number of cores in the conos object in
container$embedding has n.cores set to a relatively small value < 10.
}
