% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_enr_hmap}
\alias{get_subclust_enr_hmap}
\title{Get heatmap of subtype proportion associations for each celltype/subtype and each factor}
\usage{
get_subclust_enr_hmap(container, all_ctypes, all_res, all_factors)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{all_ctypes}{character A vector of the cell types to include}

\item{all_res}{numeric A vector of resolutions matching the all_ctypes parameter}

\item{all_factors}{numerc A vector of the factors to compute associations for}
}
\value{
A ComplexHeatmap object in container$plots$subc_enr_hmap showing the
univariate associations between cell subcluster proportions and each factor.
}
\description{
Get heatmap of subtype proportion associations for each celltype/subtype and each factor
}
