% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_plots}
\alias{get_subclust_plots}
\title{Gets cell subtype plots including an embedding, a factor association plot, and
a heatmap of differentially expressed genes between subtypes}
\usage{
get_subclust_plots(container, ctype, res, factor_use)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctype}{character The cell type for which subtypes are to be investigated}

\item{factor_use}{numeric The factor to plot scores for}

\item{resolution}{numeric The clustering resolution that was used to generate
the clustering}
}
\value{
The embedding plot for the cell type
}
\description{
Gets cell subtype plots including an embedding, a factor association plot, and
a heatmap of differentially expressed genes between subtypes
}
