% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_umap}
\alias{get_subclust_umap}
\title{Get a figure to display subclusterings at multiple resolutions}
\usage{
get_subclust_umap(container, all_ctypes, all_res, n_col = 3)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{all_ctypes}{character A vector of the cell types to include}

\item{all_res}{numeric A vector of resolutions matching the all_ctypes parameter}

\item{n_col}{numeric The number of columns to organize the figure into (default=3)}
}
\value{
the project container with the figure in container$plots$subc_umap_fig and
the individual umap plots in container$plots$subc_umaps
}
\description{
Get a figure to display subclusterings at multiple resolutions
}
