% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sums}
\alias{get_sums}
\title{calculates factor-stratified sums for each column}
\usage{
get_sums(sY, rowSel)
}
\arguments{
\item{sY}{sparse matrix Gene by cell matrix of counts}

\item{rowSel}{factor The donor that each cell is from}
}
\description{
calculates factor-stratified sums for each column
}
