% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{identify_sex_metadata}
\alias{identify_sex_metadata}
\title{Extract metadata for sex information if not provided already}
\usage{
identify_sex_metadata(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the scMinimal object with sex metadata added to the metadata
}
\description{
Extract metadata for sex information if not provided already
}
