% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{initialize_params}
\alias{initialize_params}
\title{Initialize parameters to be used throughout scITD in various functions}
\usage{
initialize_params(ctypes_use, ncores = 4, rand_seed = 10)
}
\arguments{
\item{ctypes_use}{character Names of the cell types to use for the analysis
(default=NULL)}

\item{ncores}{numeric Number of cores to use (default=4)}

\item{rand_seed}{numeric Random seed to use (default=10)}
}
\value{
A list of the experiment parameters to use.
}
\description{
Initialize parameters to be used throughout scITD in various functions
}
\examples{
param_list <- initialize_params(ctypes_use = c("CD4+ T", "CD8+ T"),
ncores = 1, rand_seed = 10)
}
