% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{merge_small_clusts}
\alias{merge_small_clusts}
\title{Merge small subclusters into larger ones}
\usage{
merge_small_clusts(con, clusts, min_cells_group)
}
\arguments{
\item{con}{conos Object for the dataset with umap projection and groups as cell types}

\item{clusts}{character The initially assigned subclusters by leiden clustering}

\item{min_cells_group}{numeric The minimum allowable cluster size}
}
\value{

}
\description{
Merge small subclusters into larger ones
}
