% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{norm_var_helper}
\alias{norm_var_helper}
\title{Calculates the normalized variance for each gene. This is based on a function
from Pagoda2}
\usage{
norm_var_helper(scMinimal)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms as well
as metadata}
}
\value{
a vector of the normalized variance for each gene
}
\description{
Calculates the normalized variance for each gene. This is based on a function
from Pagoda2
}
