% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_scMinimal.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize and log-transform count data}
\usage{
normalize_counts(count_data)
}
\arguments{
\item{count_data}{matrix or sparse matrix Gene by cell matrix of counts}
}
\value{
the normalized, log-transformed data
}
\description{
Normalize and log-transform count data
}
