% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{parse_data_by_ctypes}
\alias{parse_data_by_ctypes}
\title{Parse main counts matrix into per-celltype-matrices}
\usage{
parse_data_by_ctypes(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with scMinimal objects per cell type
}
\description{
Parse main counts matrix into per-celltype-matrices
}
