% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_sub_counts}
\alias{pbmc_sub_counts}
\title{Gene expression UMI counts matrix for all cells and all genes}
\format{
An object of class \code{dgCMatrix} with 16664 rows and 10305 columns.
}
\usage{
pbmc_sub_counts
}
\description{
Gene expression UMI counts matrix for all cells and all genes
}
\keyword{datasets}
