% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_sub_meta}
\alias{pbmc_sub_meta}
\title{Cell-wise meta data variables in a data.frame}
\format{
An object of class \code{data.frame} with 10305 rows and 3 columns.
}
\usage{
pbmc_sub_meta
}
\description{
Cell-wise meta data variables in a data.frame
}
\keyword{datasets}
