% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_sub_transformed}
\alias{pbmc_sub_transformed}
\title{Gene expression UMI counts matrix after normalizing for library size
and log transforming values}
\format{
An object of class \code{dgCMatrix} with 16664 rows and 10305 columns.
}
\usage{
pbmc_sub_transformed
}
\description{
Gene expression UMI counts matrix after normalizing for library size
and log transforming values
}
\keyword{datasets}
