% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tucker_ica.R
\name{pca_unfolded}
\alias{pca_unfolded}
\title{Computes singular-value decomposition on the unfolded tensor}
\usage{
pca_unfolded(container, ranks)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ranks}{numeric The number of factors to extract. Unlike with the Tucker
decomposition, this should be a single number.}
}
\value{
container with results of the decomposition in container$tucker_results
}
\description{
Computes singular-value decomposition on the unfolded tensor
}
