% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{plot_donor_props}
\alias{plot_donor_props}
\title{Plot donor proportions for each factor}
\usage{
plot_donor_props(donor_props, donor_scores, significance, ctype_mapping = NULL)
}
\arguments{
\item{donor_props}{data.frame Donor proportions as output from compute_donor_props()}

\item{donor_scores}{data.frame Donor scores from tucker results}

\item{significance}{numeric F-Statistics as output from compute_associations()}

\item{ctype_mapping}{character The cell types corresponding with columns of donor_props}
}
\value{
plots of donor proportions for each cell type vs donor factor scores for each factor
}
\description{
Plot donor proportions for each factor
}
