% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_donor_sig_genes}
\alias{plot_donor_sig_genes}
\title{Generate heatmap showing top genes in each cell type significantly associated
with a given factor}
\usage{
plot_donor_sig_genes(container, factor_select, top_n_per_ctype = 5)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to query}

\item{top_n_per_ctype}{numeric The number of top genes from each cell type
to plot (default=5)}
}
\value{
the project container with the plot in the slot
container$plots$donor_sig_genes$Factor#
}
\description{
Generate heatmap showing top genes in each cell type significantly associated
with a given factor
}
