% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_donor_sig_genes_v2}
\alias{plot_donor_sig_genes_v2}
\title{Generate heatmap showing top genes in each cell type significantly associated
with a given factor}
\usage{
plot_donor_sig_genes_v2(
  container,
  factor_select,
  top_n_per_ctype,
  ctypes_use = NULL,
  show_donor_labels = FALSE,
  additional_meta = NULL
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to query}

\item{top_n_per_ctype}{numeric Vector of the number of top genes from each cell type
to plot}

\item{ctypes_use}{character The cell types for which to get the top genes to make
callouts for. If NULL then uses all cell types. (default=NULL)}

\item{show_donor_labels}{logical Set to TRUE to display donor labels (default=FALSE)}

\item{additional_meta}{character Another meta variable to plot (default=NULL)}
}
\value{
the project container with the plot in the slot
container$plots$donor_sig_genes$Factor#
}
\description{
Generate heatmap showing top genes in each cell type significantly associated
with a given factor
}
