% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_donor_umap}
\alias{plot_donor_umap}
\title{Create UMAP for donor distances}
\usage{
plot_donor_umap(
  container,
  color_by_factor = NULL,
  color_by_meta = NULL,
  n_col = 1
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{color_by_factor}{numeric Number of factor to color donors by. Can be a
vector of multiple factor numbers to make several plots. (default=NULL)}

\item{color_by_meta}{character Names of meta variables to color donors by.
Can be a vector of multiple names to make several plots. (default=NULL)}

\item{n_col}{numeric The number of columns to orde the figure into (default=1)}
}
\value{
the project container with the figure in container$plots$dscores_umap
}
\description{
Create UMAP for donor distances
}
