% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_dscore_enr}
\alias{plot_dscore_enr}
\title{Compute enrichment of donor metadata categorical variables at high/low factor scores}
\usage{
plot_dscore_enr(container, factor_use, meta_var)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_use}{numeric The factor to test}

\item{meta_var}{character The name of the metadata variable to test}
}
\value{
A cowplot figure of enrichment plots.
}
\description{
Compute enrichment of donor metadata categorical variables at high/low factor scores
}
\examples{
fig <- plot_dscore_enr(test_container, factor_use=1, meta_var='lanes')
}
