% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{plot_gsea_hmap}
\alias{plot_gsea_hmap}
\title{Plot enriched gene sets from all cell types in a heatmap}
\usage{
plot_gsea_hmap(container, factor_select, thresh)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to plot}

\item{thresh}{numeric Pvalue threshold to use for including gene sets in the heatmap}
}
\value{
A stacked heatmap object from ComplexHeatmap.
}
\description{
Plot enriched gene sets from all cell types in a heatmap
}
