% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{plot_gsea_hmap_w_similarity}
\alias{plot_gsea_hmap_w_similarity}
\title{Plot enriched gene sets from all cell types in a heatmap}
\usage{
plot_gsea_hmap_w_similarity(container, factor_select, direc, thresh)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to plot}

\item{thresh}{numeric Pvalue threshold to use for including gene sets in the heatmap}

\item{set_union}{character A vector of the main pathways from collapsePathways, and it is the
union of such pathways from all cell types of the factor (default=NULL)}
}
\value{
the heatmap plot
}
\description{
Plot enriched gene sets from all cell types in a heatmap
}
