% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_loadings_annot}
\alias{plot_loadings_annot}
\title{Plot the gene x cell type loadings for a factor}
\usage{
plot_loadings_annot(
  container,
  factor_select,
  use_sig_only = FALSE,
  annot = "none",
  pathways = NULL,
  sig_thresh = 0.05,
  display_genes = FALSE,
  show_xlab = TRUE
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to plot}

\item{use_sig_only}{logical If TRUE, includes only significant genes
from jackstraw in the heatmap. If FALSE, includes all the variable genes.
(default = FALSE)}

\item{annot}{character If set to "pathways" then creates an adjacent heatmap
showing which genes are in which pathways. If set to "sig_genes" then creates
an adjacent heatmap showing which genes were significant from jackstraw. If
set to "none" no adjacent heatmap is plotted. (default="none")}

\item{pathways}{character Gene sets to plot if annot is set to "pathways"
(default=NULL)}

\item{sig_thresh}{numeric Pvalue significance threshold to use. If use_sig_only is
TRUE the threshold is used as a cutoff for genes to include. If annot is "sig_genes"
this value is used in the gene significance colormap as a minimum threshold. (default=0.05)}

\item{display_genes}{logical If TRUE, displays the names of gene names (default=FALSE)}

\item{show_xlab}{logical If TRUE, displays the xlabel 'genes' (default=TRUE)}
}
\value{
container with the plot put in container$plots$single_lds_plot
}
\description{
Plot the gene x cell type loadings for a factor
}
