% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_associations.R
\name{plot_meta_associations}
\alias{plot_meta_associations}
\title{Plot meta data associations in a heatmap}
\usage{
plot_meta_associations(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with the heatmap in the slot
container$plots$meta_associations
}
\description{
Plot meta data associations in a heatmap
}
