% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{plot_rec_errors_bar_tucker}
\alias{plot_rec_errors_bar_tucker}
\title{Plot reconstruction errors as bar plot for tucker method}
\usage{
plot_rec_errors_bar_tucker(real, shuffled)
}
\arguments{
\item{real}{list The real reconstruction errors}

\item{shuffled}{list The reconstruction errors under null model}
}
\value{
plot
}
\description{
Plot reconstruction errors as bar plot for tucker method
}
