% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{plot_scores_by_meta}
\alias{plot_scores_by_meta}
\title{Plot dotplots for each factor to compare donor scores between meta groups}
\usage{
plot_scores_by_meta(container, meta_var)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{meta_var}{character The meta data variable to compare groups for}
}
\value{
a figure of comparison plots (one for each factor) placed in
container$plots$indv_meta_scores_associations
}
\description{
Plot dotplots for each factor to compare donor scores between meta groups
}
