% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stability.R
\name{plot_stability_results}
\alias{plot_stability_results}
\title{Generate a plot for either the donor scores or loadings stability test}
\usage{
plot_stability_results(container, plt_data)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{plt_data}{character Either 'lds' or 'dsc' and indicates which plot to make}
}
\value{
the plot
}
\description{
Generate a plot for either the donor scores or loadings stability test
}
