% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{plot_subclust_associations}
\alias{plot_subclust_associations}
\title{Get plot for associations between subcluster proportions for each major cell
type and each factor}
\usage{
plot_subclust_associations(res, n_col = 2)
}
\arguments{
\item{res}{data.frame Regression statistics for each subcluster analysis}

\item{n_col}{numeric The number of columns to organize the plots into (default=2)}
}
\value{
plots of regression statistics for each subtypes at varying clustering
resolutions and for each factor
}
\description{
Get plot for associations between subcluster proportions for each major cell
type and each factor
}
