% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{plot_subclusts_only}
\alias{plot_subclusts_only}
\title{Title}
\usage{
plot_subclusts_only(
  container,
  ctype,
  resolution,
  use_scores = FALSE,
  factor_use = NULL
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctype}{character The cell type for which subtypes are to be investigated}

\item{resolution}{numeric The clustering resolution that was used to generate
the clustering}

\item{use_scores}{logical Set to TRUE to plot donor scores of a factor from 
tucker results as colors on the umap (default=FALSE)}

\item{factor_use}{numeric The factor to plot scores of if use_scores is TRUE
(default=NULL)}
}
\value{
The embedding plot for the cell type
}
\description{
Title
}
