% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{plot_var_scale_power}
\alias{plot_var_scale_power}
\title{Plot the variance scaling effect on extraction of sex-linked factor}
\usage{
plot_var_scale_power(power_results)
}
\arguments{
\item{power_results}{list F-Statistics for max factor-sex association
for Tucker decompositions over a range of variance scaling parameters}
}
\value{
plot
}
\description{
Plot the variance scaling effect on extraction of sex-linked factor
}
