% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tucker_ica.R
\name{project_new_data}
\alias{project_new_data}
\title{Project multicellular patterns to get scores on new data}
\usage{
project_new_data(new_container, old_container)
}
\arguments{
\item{new_container}{environment A project container with new data to project scores
for. The form_tensor() function should be run.}

\item{old_container}{environment The original project container that has the multicellular
gene expression patterns already extracted. These patterns will be projected onto the new data.}
}
\value{
The new container environment object with projected scores in new_container$projected_scores.
The factors will be ordered the same as the factors in old_container.
}
\description{
Project multicellular patterns to get scores on new data
}
