% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctype_vargenes.R
\name{reduce_to_vargenes}
\alias{reduce_to_vargenes}
\title{Reduce each cell type's expression matrix to just the significantly variable genes}
\usage{
reduce_to_vargenes(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with gene by cell matrices in the scMinimal_ctype
slots reduced to contain only the genes that were found to be significantly
variable across donors in at least one cell type
}
\description{
Reduce each cell type's expression matrix to just the significantly variable genes
}
