% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{render_all_lds_plots}
\alias{render_all_lds_plots}
\title{Creates a figure of all loadings plots side by side}
\usage{
render_all_lds_plots(container, n_rows)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{n_rows}{numeric The number of rows to fit the plots onto}
}
\description{
Creates a figure of all loadings plots side by side
}
