% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{render_multi_plots}
\alias{render_multi_plots}
\title{Creates a figure of all loadings plots arranged}
\usage{
render_multi_plots(container, n_rows, data_type)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{n_rows}{numeric The number of rows to fit the plots onto}

\item{data_type}{character Can be either "loadings", "gsea", or "dgenes". This
determines which list of heatmaps to organize into the figure.}
}
\description{
Creates a figure of all loadings plots arranged
}
