% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{render_subtype_plots}
\alias{render_subtype_plots}
\title{Render a figure of all subcluster plots}
\usage{
render_subtype_plots(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\description{
Render a figure of all subcluster plots
}
