% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{replace_groups_with_subclusts}
\alias{replace_groups_with_subclusts}
\title{Replace leiden groups in con object with subclusters identified for one cell type}
\usage{
replace_groups_with_subclusts(con, subclusts, sub_ctype)
}
\arguments{
\item{con}{conos Object for the dataset with umap projection and groups as cell types}

\item{subclusts}{numeric Named vector of subcluster assignments for one major cell type
only. Names should be the cell barcodes.}

\item{sub_ctype}{character The name of the major cell type that subclusts belong to}
}
\value{
a factor of the new groups with subclusters included
}
\description{
Replace leiden groups in con object with subclusters identified for one cell type
}
