% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{reshape_loadings}
\alias{reshape_loadings}
\title{Reshape loadings for a factor from linearized to matrix form}
\usage{
reshape_loadings(ldngs_row, genes, ctypes)
}
\arguments{
\item{ldngs_row}{numeric A vector of loadings values for one factor}

\item{genes}{character The gene identifiers corresponding to each loading}

\item{ctypes}{character The cell type corresponding to each loading}
}
\value{
A loadings matrix with dimensions of genes by cell types.
}
\description{
Reshape loadings for a factor from linearized to matrix form
}
