% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea_all_factors}
\alias{run_gsea_all_factors}
\title{Run gsea for all cell types and all factors and generate venn diagrams to
visualize overlap between enriched genes sets between different cell types}
\usage{
run_gsea_all_factors(
  container,
  method = "fgsea",
  thresh = 0.05,
  db_use = "GO",
  num_iter = 10000
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{method}{character The method of gsea to use. Can either be "fgsea" or
"hypergeometric". (default="fgsea")}

\item{thresh}{numeric Pvalue significance threshold (default=0.05)}

\item{db_use}{character The database of gene sets to use. Database
options include "GO", "Reactome", "KEGG", and "BioCarta". More than
one database can be used. (default="GO")}

\item{num_iter}{numeric The number of random shufflings to perform. Only
applies if using fgsea method (default=10000)}
}
\value{
project container with venn diagrams for each factor added in
container$plots slots and gsea results added in container$gsea_results
}
\description{
Run gsea for all cell types and all factors and generate venn diagrams to
visualize overlap between enriched genes sets between different cell types
}
