% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{run_subcluster_de}
\alias{run_subcluster_de}
\title{Run differential expression with cell subtypes included}
\usage{
run_subcluster_de(container, ctype, resolution)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctype}{character The cell type for which subtypes are to be investigated}

\item{resolution}{numeric The clustering resolution that was used to generate
the clustering}
}
\value{
the project container with the DE results in 
container$subcluster_de$<ctype>$<resolution>
}
\description{
Run differential expression with cell subtypes included
}
