% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tucker_ica.R
\name{run_tucker_ica}
\alias{run_tucker_ica}
\title{Run Tucker decomposition and apply ICA transformation}
\usage{
run_tucker_ica(container, ranks = NULL, shuffle = FALSE)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ranks}{numeric The number of donor, gene, and cell type ranks, respectively,
to decompose to using Tucker decomposition. If NULL, uses ranks in container$experiment_params
field. (default=NULL)}

\item{shuffle}{logical If TRUE, randomly shuffles cell to donor linkages, resulting in a random
tensor (default=FALSE)}
}
\value{
container with results of the decomposition in container$tucker_results
}
\description{
Run Tucker decomposition and apply ICA transformation
}
