% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw_v2.R
\name{sample_fibers_v2}
\alias{sample_fibers_v2}
\title{Get list of tensor fibers to shuffle}
\usage{
sample_fibers_v2(tensor_data, n_fibers)
}
\arguments{
\item{tensor_data}{list The tensor data including donor, gene, and cell type labels
as well as the tensor array itself}

\item{n_fibers}{numeric The number of fibers to get}
}
\value{
a list of gene and cell type indices for the randomly selected fibers
}
\description{
Get list of tensor fibers to shuffle
}
