% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{scale_fontsize}
\alias{scale_fontsize}
\title{Scale font size. From simplifyEnrichment package.
https://github.com/jokergoo/simplifyEnrichment/blob/master/R/ht_clusters.R}
\usage{
scale_fontsize(x, rg = c(1, 30), fs = c(4, 16))
}
\arguments{
\item{x}{A numeric vector.}

\item{rg}{The range.}

\item{fs}{Range of the font size.}
}
\value{
A numeric vector.
}
\description{
Scale font size. From simplifyEnrichment package.
https://github.com/jokergoo/simplifyEnrichment/blob/master/R/ht_clusters.R
}
