% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{scale_variance}
\alias{scale_variance}
\title{Scale variance across donors for each gene within each cell type. Generally, this should be done through calling the
form_tensor() wrapper function.}
\usage{
scale_variance(container, var_scale_power)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{var_scale_power}{numeric Exponent of normalized variance that is
used for variance scaling. Variance for each gene
is initially set to unit variance across donors (for a given cell type).
Variance for each gene is then scaled by multiplying the unit scaled values
by each gene's normalized variance (where the effect of the mean-variance
dependence is taken into account) to the exponent specified here.
If NULL, uses var_scale_power from container$experiment_params.}
}
\value{
The project container with the variance altered for each gene within
the pseudobulked matrices for each cell type.
}
\description{
Scale variance across donors for each gene within each cell type. Generally, this should be done through calling the
form_tensor() wrapper function.
}
