% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{set_experiment_params}
\alias{set_experiment_params}
\title{Set experiment parameters}
\usage{
set_experiment_params(
  container,
  ctypes_use = NULL,
  scale_var = NULL,
  ranks = NULL,
  rotate_modes = NULL,
  var_scale_power = NULL,
  ncores = NULL
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctypes_use}{character Names of the cell types to use for the analysis
(default=NULL)}

\item{scale_var}{logical TRUE to scale the gene expression variance across donors
for each cell type. If FALSE then all genes are scaled to unit variance across
donors for each cell type. (default=NULL)}

\item{ranks}{numeric The number of donor, gene, and cell type ranks, respectively,
to decompose to using Tucker decomposition (default=NULL)}

\item{rotate_modes}{character The names of the tensor modes to rotate with
ICA during Tucker decomposition. Can include 'donors', 'genes', and/or 'ctypes'
(default='NULL')}

\item{var_scale_power}{numeric Exponent of normalized variance that is
used for variance scaling. Variance for each gene
is initially set to unit variance across donors (for a given cell type).
Variance for each gene is then scaled by multiplying the unit scaled values
by each gene's normalized variance (where the effect of the mean-variance
dependence is taken into account) to the exponent specified here. (default=NULL)}

\item{ncores}{numeric Number of cores to use (default=NULL)}
}
\value{
the project container with updated experiment parameters in
container$experiment_params
}
\description{
Set experiment parameters
}
