% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{set_vargenes_container}
\alias{set_vargenes_container}
\title{Get union of significantly variable genes from all cell types}
\usage{
set_vargenes_container(container, vargenes)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{vargenes}{character Significantly variable genes from one cell type}
}
\value{
the project container with updated slot container$all_vargenes
}
\description{
Get union of significantly variable genes from all cell types
}
