% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{shuffle_fibers}
\alias{shuffle_fibers}
\title{Shuffle elements within selected fibers}
\usage{
shuffle_fibers(tensor_data, s_fibers)
}
\arguments{
\item{tensor_data}{list The tensor data including donor, gene, and cell type labels
as well as the tensor array itself}

\item{s_fibers}{list Gene and cell type indices for the randomly selected fibers}
}
\value{
the tensor_data with the values for the selected fibers shuffled
}
\description{
Shuffle elements within selected fibers
}
