% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{sim_helper}
\alias{sim_helper}
\title{Get data for DE genes that will correspond with the main simulation}
\usage{
sim_helper(groups, g_frac, n_de_genes_per_group, params, meta, de_strength)
}
\arguments{
\item{groups}{numeric Splatter group numbers corresponding with the groups to give the DE genes to}

\item{g_frac}{numeric Decimal indicating the fractional size of the DE groups}

\item{n_de_genes_per_group}{numeric The desired number of DE genes to get}

\item{params}{SplatParams Parameters used for the simulation}

\item{meta}{DataFrame The colData of the simulation}

\item{de_strength}{numeric An average foldchange for the DE genes}
}
\value{
normalized counts matrix of DE genes for the specified groups
}
\description{
Get data for DE genes that will correspond with the main simulation
}
