% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{stack_tensor}
\alias{stack_tensor}
\title{Create the tensor object by stacking each bulk cell type matrix}
\usage{
stack_tensor(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with the tensor data in container$tensor_data
}
\description{
Create the tensor object by stacking each bulk cell type matrix
}
