% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{stop_wrap}
\alias{stop_wrap}
\title{Helper function from simplifyEnrichment package}
\usage{
stop_wrap(...)
}
\arguments{
\item{...}{other parameters}
}
\value{
No value is returned.
}
\description{
Helper function from simplifyEnrichment package
}
