% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_container}
\alias{test_container}
\title{Tiny data container object for unit tests only}
\format{
An object of class \code{environment} of length 14.
}
\usage{
test_container
}
\description{
Tiny data container object for unit tests only
}
\keyword{datasets}
