% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Contains real tensor data and Tucker results for testing the decomposition}
\format{
An object of class \code{list} of length 2.
}
\usage{
test_df
}
\description{
Contains real tensor data and Tucker results for testing the decomposition
}
\keyword{datasets}
