% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_custom_tucker.R
\name{tucker_sparse}
\alias{tucker_sparse}
\title{Tucker Decomposition adapted from rTensor but with sparsity constraints added. This
function is still being tested, so use with caution.
https://github.com/jamesyili/rTensor/blob/master/R/rTensor_Decomp.R}
\usage{
tucker_sparse(tnsr, ranks = NULL, max_iter = 25, tol = 1e-05, sparsity = 5)
}
\arguments{
\item{tnsr}{Tensor with K modes.}

\item{ranks}{numeric Vector of the modes of the output core Tensor (default=NULL)}

\item{max_iter}{numeric Maximum number of iterations if error stays above tol (default=25)}

\item{tol}{numeric Relative Frobenius norm error tolerance (default=1e-5)}

\item{sparsity}{numeric Higher is more sparse (default=5)}
}
\value{
A list containing all the Tucker decomposition results components.
}
\description{
Tucker Decomposition adapted from rTensor but with sparsity constraints added. This
function is still being tested, so use with caution.
https://github.com/jamesyili/rTensor/blob/master/R/rTensor_Decomp.R
}
