% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{vargenes_anova}
\alias{vargenes_anova}
\title{Compute significantly variable genes via anova. Generally, this should be done through calling the
form_tensor() wrapper function.}
\usage{
vargenes_anova(scMinimal, ncores)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms}

\item{ncores}{numeric Number of cores to use}
}
\value{
A list of raw p-values for each gene.
}
\description{
Compute significantly variable genes via anova. Generally, this should be done through calling the
form_tensor() wrapper function.
}
