% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vargenes.R
\name{vargenes_shuffle}
\alias{vargenes_shuffle}
\title{Compute significant variable genes via empirical shuffling}
\usage{
vargenes_shuffle(scMinimal, num_iter, ncores)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms}

\item{num_iter}{numeric Number of null statistics to generate}

\item{ncores}{numeric Number of cores to use}
}
\value{
the raw pvalues for each gene
}
\description{
Compute significant variable genes via empirical shuffling
}
