% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data to remove genes only expressed in a few cells and donors
with very few cells}
\usage{
clean_data(scMinimal, donor_min_cells = 5, gene_min_cells = 5)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms}

\item{donor_min_cells}{numeric Minimum threshold for number of cells per
donor (default=5)}

\item{gene_min_cells}{numeric Minimum threshold for number of cells
with nonzero expression of a gene (default=5)}
}
\value{
an scMinimal environment with cleaned counts matrix, transformed
counts matrix, and metadata fields
}
\description{
Clean data to remove genes only expressed in a few cells and donors
with very few cells
}
