% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colMeanVars}
\alias{colMeanVars}
\title{calculates column mean and variance, optionally taking a subset of rows to operate on}
\usage{
colMeanVars(sY, rowSel, ncores = 1L)
}
\arguments{
\item{sY}{sparse matrix Gene by cell matrix of counts}

\item{rowSel}{numeric The selected rows (genes)}

\item{ncores}{numeric The number of cores}
}
\description{
calculates column mean and variance, optionally taking a subset of rows to operate on
}
\examples{
library(Matrix)
donor_by_gene <- rbind(c(9,2,1,5), c(3,3,1,2))
donor_by_gene <- Matrix(donor_by_gene, sparse = TRUE)
result <- colMeanVars(donor_by_gene, rowSel = NULL, ncores=1)

}
