% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gn.R
\name{convert_gn}
\alias{convert_gn}
\title{Convert gene names}
\usage{
convert_gn(container, genes)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{genes}{character Vector of the gene identifiers to be converted to
gene symbols}
}
\value{
character vector of corresponding gene symbols
}
\description{
Convert gene names
}
