% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctype_vargenes.R
\name{get_ctype_data}
\alias{get_ctype_data}
\title{Create an environment for each cell type}
\usage{
get_ctype_data(container, make_clean = TRUE)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{make_clean}{logical TRUE to apply minimum thresholds for number of cells
expressing a gene and number of cells per donor (default=TRUE)}
}
\value{
the project container with the scMinimal environments added into
the container$scMinimal_ctypes slot
}
\description{
Create an environment for each cell type
}
