% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctype_vargenes.R
\name{get_ctype_vargenes}
\alias{get_ctype_vargenes}
\title{Partition main gene by cell matrix into per cell type matrices with significantly
variable genes only}
\usage{
get_ctype_vargenes(container, method = "anova", thresh = 0.01)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{method}{character The method used to select significantly variable
genes across donors within a cell type (default="anova")}

\item{thresh}{numeric Pvalue threshold to use for gene significance
(default=0.01)}
}
\value{
the project container with scMinimal environments added for each
cell type. The scMinimal environments contain expression count matrices
with columns as the cells belonging to the respective cell type and rows
as the genes which were identified as significantly variable across donors
in any cell type. This is in the "count_data_sparse" slot. The
normalized, log-transformed count matrices with the same dimensions are
also available in the "data_sparse slot". Metadata is also present
in these sub-containers for the corresponding included cells and is in the
"metadata" slot.
}
\description{
Partition main gene by cell matrix into per cell type matrices with significantly
variable genes only
}
