% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{get_lm_pvals}
\alias{get_lm_pvals}
\title{Compute gene-factor associations}
\usage{
get_lm_pvals(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with the adjust p-values for the gene-factor
associations in container$gene_score_associations
}
\description{
Compute gene-factor associations
}
\examples{
test_container <- get_lm_pvals(test_container)
}
