% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{get_min_sig_genes}
\alias{get_min_sig_genes}
\title{Evaluate the minimum number for significant genes in any factor for a given number of
donor factors extracted by tucker.}
\usage{
get_min_sig_genes(container, donor_rank_range, thresh = 0.05)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses. Should have}

\item{donor_rank_range}{numeric Range of possible number of donor factors to use.}

\item{thresh}{numeric Pvalue threshold for significant genes in calculating the
number of significant genes identified per factor. (default=0.05)}
}
\value{
the project container with a plot of the minimum significant genes for
each decomposition with varying number of donor factors located in
container$plots$min_sig_genes
}
\description{
Evaluate the minimum number for significant genes in any factor for a given number of
donor factors extracted by tucker.
}
