% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{get_normalized_variance}
\alias{get_normalized_variance}
\title{Get normalized variance for each gene, taking into account mean-variance trend}
\usage{
get_normalized_variance(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
The project container with vectors of normalized variances values in scMinimal objects
for each cell type. Generally, this should be done through calling the
form_tensor() wrapper function.
}
\description{
Get normalized variance for each gene, taking into account mean-variance trend
}
