% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{get_pseudobulk}
\alias{get_pseudobulk}
\title{Collapse data from cell-level to donor-level via summing counts}
\usage{
get_pseudobulk(container, shuffle = FALSE, shuffle_within = NULL)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{shuffle}{logical Set to TRUE to shuffle cell-donor linkages (default=FALSE)}

\item{shuffle_within}{character A metadata variable to shuffle cell-donor linkages
within (default=NULL)}
}
\value{
the project container with pseudobulked count matrices in
scMinimal_ctype$ctype$pseudobulk slots for each cell type
}
\description{
Collapse data from cell-level to donor-level via summing counts
}
