% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tucker.R
\name{get_significance_vectors}
\alias{get_significance_vectors}
\title{Get vectors indicating which genes are significant in which cell types
for a factor of interest}
\usage{
get_significance_vectors(container, factor_select, ctypes)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to query}

\item{ctypes}{character The cell types used in all the analysis ordered
as they appear in the loadings matrix}
}
\value{
A list of the adjusted p-values for expression of each gene in each cell type in
association with a factor of interest.
}
\description{
Get vectors indicating which genes are significant in which cell types
for a factor of interest
}
