% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_data.R
\name{get_sim_data}
\alias{get_sim_data}
\title{Generate simulated data with subgroups of donors expressing multi-cell type processes}
\usage{
get_sim_data(
  donors_total,
  cells_per_donor,
  n_processes,
  donors_per_process,
  n_ctypes = 2,
  de_prob = 0.05,
  de_strength = 2
)
}
\arguments{
\item{donors_total}{numeric Total number of donors to include in the simulation}

\item{cells_per_donor}{numeric Average number of cells per donor to generate. Total
dataset size will be cells_per_donor * donors_total number of cells.}

\item{n_processes}{numeric The number of multi-cell type processes/factors to generate.
Each process will be overexpressed in a subgroup of donors.}

\item{donors_per_process}{numeric The size of each subgroup of donors to go along with
each multi-cell type process. donors_per_process * n_processes should not exceed
donors_total.}

\item{n_ctypes}{numeric The number of different cell types to be generated (default=2)}

\item{de_prob}{numeric The probability that each gene gets upregulated in each cell
type in each process (default=.05)}

\item{de_strength}{numeric This is roughly the average effect size for all differentially
expressed genes used to generate the distinc processes (default=2)}
}
\value{
normalized gene by cell count matrix, raw gene by cell count matrix,
metadata data.frame, and matrix of differentially expressed genes
}
\description{
Generate simulated data with subgroups of donors expressing multi-cell type processes
}
\examples{
get_sim_data(donors_total=50, cells_per_donor=200, n_processes=2, donors_per_process=10,
n_ctypes=2, de_prob=0.05, de_strength=2)
}
