% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subclust_enr_dotplot}
\alias{get_subclust_enr_dotplot}
\title{Get scatter plot for association of a cell subtype proportion with scores for a factor}
\usage{
get_subclust_enr_dotplot(
  container,
  ctype,
  res,
  subtype,
  factor_use,
  ctype_cur = ctype
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ctype}{character The cell type to plot}

\item{res}{numeric The subcluster resolution to use}

\item{subtype}{numeric The number corresponding with the subtype of the major
cell type to plot}

\item{factor_use}{numeric The factor to plot}

\item{ctype_cur}{character The name of the major cell type used in the main analysis}
}
\value{
A ggplot object of each donor's cell subcluster proportions against donor
scores for a selected factor.
}
\description{
Get scatter plot for association of a cell subtype proportion with scores for a factor
}
