% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{get_subtype_prop_associations}
\alias{get_subtype_prop_associations}
\title{Compute associations between donor factor scores and donor proportions of cell subtypes}
\usage{
get_subtype_prop_associations(
  container,
  max_res,
  stat_type,
  integration_var = NULL,
  min_cells_group = 50,
  n_col = 2
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{max_res}{numeric The maximum clustering resolution to use}

\item{stat_type}{character Either "fstat" to get F-Statistics, "adj_rsq" to get adjusted
R-squared values, or "adj_pval" to get adjusted pvalues.}

\item{integration_var}{character The meta data variable to use for creating
the joint embedding with Conos if not already provided in container$embedding (default=NULL)}

\item{min_cells_group}{numeric The minimum allowable size for cell subpopulations
(default=50)}

\item{n_col}{numeric The number of columns to organize the plots into (default=2)}
}
\value{
the project container with a plot of association results in
container$plots$subtype_prop_factor_associations
}
\description{
Compute associations between donor factor scores and donor proportions of cell subtypes
}
