% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sums}
\alias{get_sums}
\title{Calculates factor-stratified sums for each column. Adapted from pagoda2.
https://github.com/kharchenkolab/pagoda2/blob/main/src/misc2.cpp}
\usage{
get_sums(sY, rowSel)
}
\arguments{
\item{sY}{sparse matrix Gene by cell matrix of counts}

\item{rowSel}{factor The donor that each cell is from}
}
\value{
matrix of summed counts per gene per sample
}
\description{
Calculates factor-stratified sums for each column. Adapted from pagoda2.
https://github.com/kharchenkolab/pagoda2/blob/main/src/misc2.cpp
}
