% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_scMinimal.R
\name{identify_sex_metadata}
\alias{identify_sex_metadata}
\title{Extract metadata for sex information if not provided already}
\usage{
identify_sex_metadata(scMinimal)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms as well
as metadata}
}
\value{
the scMinimal object with sex metadata added to the metadata
}
\description{
Extract metadata for sex information if not provided already
}
