% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{initialize_params}
\alias{initialize_params}
\title{Update any of the experiment parameters}
\usage{
initialize_params(ctypes_use, ncores = 4, rand_seed = 10)
}
\arguments{
\item{ctypes_use}{character Names of the cell types to use for the analysis
(default=NULL)}

\item{ncores}{numeric Number of cores to use (default=4)}

\item{rand_seed}{numeric Random seed to use (default=10)}
}
\value{
a list of the experiment parameters to use
}
\description{
Update any of the experiment parameters
}
