% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_scMinimal.R
\name{instantiate_scMinimal}
\alias{instantiate_scMinimal}
\title{Create an scMinimal object}
\usage{
instantiate_scMinimal(data_sparse, count_data, meta_data)
}
\arguments{
\item{data_sparse}{sparseMatrix Matrix of normalized, log-transformed
counts with genes as rows and cells as columns}

\item{count_data}{sparseMatrix Matrix of raw counts with genes as rows
and cells as columns}

\item{meta_data}{data.frame Metadata with cells as rows and variables
as columns. Number of rows in metadata should equal number of columns
in data_sparse}
}
\value{
scMinimal object
}
\description{
Create an scMinimal object
}
