% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_container.R
\name{make_new_container}
\alias{make_new_container}
\title{Create a container to store all data and results for the project}
\usage{
make_new_container(
  scMinimal,
  ctypes_use = NULL,
  gn_convert = NULL,
  scale_var = TRUE,
  var_scale_power = NULL,
  tucker_type = "regular",
  rotation_type = "ica",
  rotate_modes = "donors",
  ranks = NULL,
  ncores = 4,
  rand_seed = 10
)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms as well
as metadata}

\item{ctypes_use}{character Names of the cell types to use for the analysis
(default=NULL)}

\item{gn_convert}{data.frame Gene identifier -> gene name conversions table.
Gene identifiers used in counts matrices should appear in the first column and
the corresponding gene symbols should appear in the second column. Can remain
NULL if the identifiers are already gene symbols. (default=NULL)}

\item{scale_var}{logical TRUE to scale the gene expression variance across donors
for each cell type. If FALSE then all genes are scaled to unit variance across
donors for each cell type. (default=TRUE)}

\item{var_scale_power}{numeric Exponent of normalized variance that is
used for variance scaling. Variance for each gene
is initially set to unit variance across donors (for a given cell type).
Variance for each gene is then scaled by multiplying the unit scaled values
by each gene's normalized variance (where the effect of the mean-variance
dependence is taken into account) to the exponent specified here. (default=NULL)}

\item{tucker_type}{character Set to 'regular' to run regular tucker or to 'sparse' to run tucker
with sparsity constraints (default='regular')}

\item{rotation_type}{character Set to 'ica' to perform ICA rotation on resulting donor factor
matrix and loadings. Otherwise set to 'varimax' to perform varimax rotation. (default='ica')}

\item{rotate_modes}{character The names of the tensor modes to rotate with
ICA during Tucker decomposition. Can include 'donors', 'genes', and/or 'ctypes'
(default='donors')}

\item{ranks}{numeric The number of donor, gene, and cell type ranks, respectively,
to decompose to using Tucker decomposition (default=NULL)}

\item{ncores}{numeric Number of cores to use (default=4)}

\item{rand_seed}{numeric Random seed to use for all stochastic analyses to make
results reproducible (default=10)}
}
\value{
project container that stores sub-containers
for each cell type as well as results and plots from all analyses
}
\description{
Create a container to store all data and results for the project
}
