% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Helper function to normalize and log-transform count data}
\usage{
normalize_counts(count_data, scale_factor = 10000)
}
\arguments{
\item{count_data}{matrix or sparse matrix Gene by cell matrix of counts}

\item{scale_factor}{numeric The number that gets multiplied by fractional counts
during normalization of the pseudobulked data (default=10000)}
}
\value{
the normalized, log-transformed data
}
\description{
Helper function to normalize and log-transform count data
}
