% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{optimize_var_scale_power}
\alias{optimize_var_scale_power}
\title{Evaluate ability to extract sex-linked factor over varying the
variance scaling exponent parameter}
\usage{
optimize_var_scale_power(container, max_ranks_test)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{max_ranks_test}{numeric Vector of length 3 with maximum number of
ranks to test for donor, gene, and cell type modes in that order}
}
\value{
the experiment container with added plot and results
}
\description{
Evaluate ability to extract sex-linked factor over varying the
variance scaling exponent parameter
}
