% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_ranks_tucker.R
\name{optimize_var_scale_power}
\alias{optimize_var_scale_power}
\title{Evaluate ability to extract sex-linked factor over varying the
variance scaling exponent parameter}
\usage{
optimize_var_scale_power(
  container,
  min_ranks_test,
  max_ranks_test,
  min_power_test,
  max_power_test,
  tucker_type = "regular",
  rotation_type = "ica",
  norm_method = "trim",
  scale_factor = 10000,
  batch_var = NULL
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{min_ranks_test}{numeric Vector of length 3 with minimum number of
ranks to test for donor, gene, and cell type modes in that order}

\item{max_ranks_test}{numeric Vector of length 3 with maximum number of
ranks to test for donor, gene, and cell type modes in that order}

\item{min_power_test}{numeric The minimum exponent parameter value to test}

\item{max_power_test}{numeric The maximum parameter value to test}

\item{tucker_type}{character Set to 'regular' to run regular tucker or to 'sparse' to run tucker
with sparsity constraints (default='regular')}

\item{rotation_type}{character Set to 'ica' to perform ICA rotation on resulting donor factor
matrix and loadings. Otherwise set to 'varimax' to perform varimax rotation. (default='ica')}

\item{norm_method}{character The normalization method to use on the pseudobulked
count data. Set to 'regular' to do standard normalization of dividing by
library size. Set to 'trim' to use edgeR trim-mean normalization, whereby counts
are divided by library size times a normalization factor. (default='trim')}

\item{scale_factor}{numeric The number that gets multiplied by fractional counts
during normalization of the pseudobulked data (default=10000)}

\item{batch_var}{character A batch variable from metadata to remove. No batch
correction applied if NULL. (default=NULL)}
}
\value{
the experiment container with plot in container$plots$var_scale_plot
}
\description{
Evaluate ability to extract sex-linked factor over varying the
variance scaling exponent parameter
}
