% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{plot_gsea_hmap}
\alias{plot_gsea_hmap}
\title{Plot enriched gene sets from all cell types in a heatmap}
\usage{
plot_gsea_hmap(up_down_sets, thresh)
}
\arguments{
\item{up_down_sets}{list The gsea results with either only the up genes sets or down gene sets}

\item{thresh}{numeric Pvalue threshold to use for including gene sets in the heatmap}
}
\value{
the heatmap plot
}
\description{
Plot enriched gene sets from all cell types in a heatmap
}
