% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{plot_gsea_hmap_w_similarity}
\alias{plot_gsea_hmap_w_similarity}
\title{Plot already computed enriched gene sets to show semantic similarity between sets}
\usage{
plot_gsea_hmap_w_similarity(
  container,
  factor_select,
  direc,
  thresh,
  exclude_words = character(0)
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to plot}

\item{direc}{character Set to either 'up' or 'down' to use the appropriate sets}

\item{thresh}{numeric Pvalue threshold to use for including gene sets in the heatmap}

\item{exclude_words}{character Vector of words to exclude from word cloud
(default=character(0))}
}
\value{
No value is returned. A heatmap showing enriched gene sets clustered by
semantic similarity is drawn.
}
\description{
Plot already computed enriched gene sets to show semantic similarity between sets
}
