% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{plot_gsea_hmap_w_similarity}
\alias{plot_gsea_hmap_w_similarity}
\title{Plot already computed enriched gene sets to show semantic similarity between sets}
\usage{
plot_gsea_hmap_w_similarity(container, factor_select, direc, thresh)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor to plot}

\item{direc}{character Set to either 'up' or 'down' to use the appropriate sets}

\item{thresh}{numeric Pvalue threshold to use for including gene sets in the heatmap}
}
\value{
the heatmap plot
}
\description{
Plot already computed enriched gene sets to show semantic similarity between sets
}
