% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_analysis.R
\name{plot_stability_results}
\alias{plot_stability_results}
\title{Plot stability results}
\usage{
plot_stability_results(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the plot of stability results
}
\description{
Plot stability results
}
