% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{plot_subclust_associations}
\alias{plot_subclust_associations}
\title{Plot association significances for varying clustering resolutions}
\usage{
plot_subclust_associations(res, n_col = 2)
}
\arguments{
\item{res}{data.frame Regression statistics for each subcluster analysis}

\item{n_col}{numeric The number of columns to organize the plots into (default=2)}
}
\value{
A cowplot of ggplot objects showing statistics for regressions of proportions of
each cell subtype (at varying clustering resolutions) against each factor.
}
\description{
Plot association significances for varying clustering resolutions
}
