% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{Gets a conos object of the data, aligning datasets across a specified variable such as
batch or donors. This can be run independently or through get_subtype_prop_associations().}
\usage{
reduce_dimensions(
  container,
  integration_var,
  ncores = container$experiment_params$ncores
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{integration_var}{character The meta data variable to use for creating
the joint embedding with Conos.}

\item{ncores}{numeric The number of cores to use (default=container$experiment_params$ncores)}
}
\value{
The project container with a conos object in container$embedding.
}
\description{
Gets a conos object of the data, aligning datasets across a specified variable such as
batch or donors. This can be run independently or through get_subtype_prop_associations().
}
