% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prop_associations.R
\name{reduce_dimensions}
\alias{reduce_dimensions}
\title{Gets umap coordinates if not already provided in container$scMinimal_full$umap}
\usage{
reduce_dimensions(scMinimal)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms as well
as metadata}
}
\value{
a dataframe with umap coordinates of each cell in the dataset
}
\description{
Gets umap coordinates if not already provided in container$scMinimal_full$umap
}
