% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{reduce_to_vargenes}
\alias{reduce_to_vargenes}
\title{Reduce each cell type's expression matrix to just the significantly variable genes}
\usage{
reduce_to_vargenes(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
the project container with pseudobulked matrices reduced to only
the most variable genes
}
\description{
Reduce each cell type's expression matrix to just the significantly variable genes
}
