% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Compute enriched gene sets in a cell type for a factor using fgsea}
\usage{
run_fgsea(
  container,
  factor_select,
  ctype,
  db_use = "GO",
  num_iter = 10000,
  print_res = T
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor of interest}

\item{ctype}{character The cell type of interest}

\item{db_use}{character The database of gene sets to use. Database
options include "GO", "Reactome", "KEGG", and "BioCarta". More than
one database can be used. (default="GO")}

\item{num_iter}{numeric The number of random shufflings to perform
(default=10000)}

\item{print_res}{logical If TRUE, prints out the top up/down gene sets
(default=TRUE)}
}
\value{
data.frame of the fgsea results limited to genes passing the
significance threshold.
}
\description{
Compute enriched gene sets in a cell type for a factor using fgsea
}
