% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run fgsea for one cell type of one factor. Note, this uses}
\usage{
run_fgsea(container, factor_select, ctype, db_use = "GO")
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor of interest}

\item{ctype}{character The cell type of interest}

\item{db_use}{character The database of gene sets to use. Database
options include "GO", "Reactome", "KEGG", "BioCarta", and "Hallmark". More than
one database can be used. (default="GO")}
}
\value{
data.frame of the fgsea results (including non-significant results)
}
\description{
Run fgsea for one cell type of one factor. Note, this uses
}
