% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea_one_factor}
\alias{run_gsea_one_factor}
\title{Run gsea separately for all cell types of one specified factor}
\usage{
run_gsea_one_factor(
  container,
  factor_select,
  method = "fgsea",
  thresh = 0.05,
  db_use = "GO"
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{factor_select}{numeric The factor of interest}

\item{method}{character The method of gsea to use. Can either be "fgsea" or
"hypergeometric". (default="fgsea")}

\item{thresh}{numeric Pvalue significance threshold to use. Will include gene sets in
resulting heatmap if pvalue is below this threshold for at least one cell type. (default=0.05)}

\item{db_use}{character The database of gene sets to use. Database
options include "GO", "Reactome", "KEGG", and "BioCarta". More than
one database can be used. (default="GO")}
}
\value{
a heatmap plot of the gsea results
}
\description{
Run gsea separately for all cell types of one specified factor
}
