% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jackstraw.R
\name{run_jackstraw}
\alias{run_jackstraw}
\title{Run tensor-based jackstraw to get gene_cell type combinations that are significantly
associated with donor scores for factors extracted by Tucker decomposition}
\usage{
run_jackstraw(
  container,
  ranks,
  n_fibers = 100,
  n_iter = 500,
  tucker_type = "regular",
  rotation_type = "hybrid"
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ranks}{numeric The number of donor ranks and gene ranks to decompose to
using Tucker decomposition}

\item{n_fibers}{numeric The number of fibers the randomly shuffle in each iteration
(default=100)}

\item{n_iter}{numeric The number of shuffling iterations to complete (default=500)}

\item{tucker_type}{character Set to 'regular' to run regular tucker or to 'sparse' to run tucker
with sparsity constraints (default='regular')}

\item{rotation_type}{character Set to 'hybrid' to perform hybrid rotation on resulting donor factor
matrix and loadings. Otherwise set to 'ica_lds' to perform ica rotation on loadings or
ica_dsc to perform ica on donor scores. (default='hybrid')}
}
\value{
the project container with adjusted pvalues in container$gene_score_associations
}
\description{
Run tensor-based jackstraw to get gene_cell type combinations that are significantly
associated with donor scores for factors extracted by Tucker decomposition
}
