% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_analysis.R
\name{run_stability_analysis}
\alias{run_stability_analysis}
\title{Compute stability of Tucker decomposition by comparing results to those computed
on the original dataset downsampled randomly}
\usage{
run_stability_analysis(container, downsample_ratio = 0.9, n_iter = 500)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{downsample_ratio}{numeric A decimal between 0-1 that indicates the fraction
of cells to retain in the subsampled datasets (default=0.9)}

\item{n_iter}{numeric The number of downsampling iterations to run (default=500)}
}
\value{
the container with a plot of the results in container$plots$stability_plot
and the raw results in container$stability_results
}
\description{
Compute stability of Tucker decomposition by comparing results to those computed
on the original dataset downsampled randomly
}
