% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tucker_ica.R
\name{run_tucker_ica}
\alias{run_tucker_ica}
\title{Run Tucker decomposition and apply ICA transformation}
\usage{
run_tucker_ica(
  container,
  ranks,
  tucker_type = "regular",
  rotation_type = "ica",
  sparsity = 5
)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}

\item{ranks}{numeric The number of donor, gene, and cell type ranks, respectively,
to decompose to using Tucker decomposition.}

\item{tucker_type}{character Set to 'regular' to run regular tucker or to 'sparse' to run tucker
with sparsity constraints (default='regular')}

\item{rotation_type}{character Set to 'ica' to perform ICA rotation on resulting donor factor
matrix and loadings. Otherwise set to 'varimax' to perform varimax rotation. (default='ica')}

\item{sparsity}{numeric To use with sparse tucker. Higher indicates more sparse (default=5)}
}
\value{
container with results of the decomposition in container$tucker_results
}
\description{
Run Tucker decomposition and apply ICA transformation
}
