% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_tensor.R
\name{stack_tensor}
\alias{stack_tensor}
\title{Create the tensor object by stacking each pseudobulk cell type matrix. 
Generally, this should be done through calling the form_tensor() wrapper function.}
\usage{
stack_tensor(container)
}
\arguments{
\item{container}{environment Project container that stores sub-containers
for each cell type as well as results and plots from all analyses}
}
\value{
The project container with the list of tensor data in container$tensor_data.
}
\description{
Create the tensor object by stacking each pseudobulk cell type matrix. 
Generally, this should be done through calling the form_tensor() wrapper function.
}
