% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{List of dfs for testing ica}
\format{
An object of class \code{list} of length 2.
}
\usage{
test_df
}
\description{
List of dfs for testing ica
}
\keyword{datasets}
