% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vargenes.R
\name{vargenes_anova}
\alias{vargenes_anova}
\title{Compute significantly variable genes via anova}
\usage{
vargenes_anova(scMinimal, ncores)
}
\arguments{
\item{scMinimal}{environment A sub-container for the project typically
consisting of gene expression data in its raw and processed forms}

\item{ncores}{numeric Number of cores to use}
}
\value{
the raw pvalues for each gene
}
\description{
Compute significantly variable genes via anova
}
